feat: Support version 2 of projectconfigs endpoint
In the first version of projectconfigs endpoint projects were identified
by a numeric ID. Since the second version, they are instead identified by
the public key. This means we have to look at projects a bit differently.

The bulk of the change is actually telling load generator to use different
public keys for different projects. The key in the config is assumed to be
a hex number (a long one, sure) and the project ID is added (mathematically)
to it, then it goes back to hex.